my_ngo_project/requirements.txt
Django==5.0.6
django-allauth==0.61.1
django-ckeditor-5==0.2.12
requests==2.31.0
urllib3==2.2.1
Pillow==10.2.0